Reorganize screen structure
